RuleCellFactory takes only live cells into account

+++ src/main/java/be/cegeka/gameoflife/domain/rules/RuleCell.java
@@ -3,9 +3,9 @@ package be.cegeka.gameoflife.domain.rules;
 public class RuleCell {
 
     private final boolean alive;
-    private final int amountOfLiveNeighbours;
+    private final long amountOfLiveNeighbours;
 
-    RuleCell(boolean alive, int amountOfLiveNeighbours) {
+    RuleCell(boolean alive, long amountOfLiveNeighbours) {
         this.alive = alive;
         this.amountOfLiveNeighbours = amountOfLiveNeighbours;
     }
@@ -18,7 +18,7 @@ public class RuleCell {
         return !isAlive();
     }
 
-    int amountOfLiveNeighbours() {
+    long amountOfLiveNeighbours() {
         return amountOfLiveNeighbours;
     }
 }
